+++ src/modules/library/ERC1404InterfaceId.sol
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: MPL-2.0
+
+pragma solidity ^0.8.20;
+
+/**
+ * @title ERC1404InterfaceId
+ * @dev ERC-165 interface IDs for ERC-1404 interfaces.
+ */
+library ERC1404InterfaceId {
+    bytes4 public constant IERC1404_INTERFACE_ID = 0xab84a5c8;
+}