+++ test/RuleWhitelist/CMTATIntegrationBase.sol
@@ -6,7 +6,6 @@ import {CMTATStandalone} from "CMTAT/deployment/CMTATStandalone.sol";
 // forge-lint: disable-next-line(unaliased-plain-import)
 import "../HelperContract.sol";
 
-
 /**
  * @title Base integration test with the CMTAT (whitelist)
  */